fix(Steps): remove externalClasses and style props
