+++ packages/products/tdesign-miniprogram/src/steps/type.ts
@@ -28,13 +28,6 @@ export interface TdStepsProps {
     value?: 'default' | 'process' | 'finish' | 'error';
   };
   /**
-   * 组件类名，用于设置组件外层元素元素类名
-   */
-  externalClasses?: {
-    type: ArrayConstructor;
-    value?: ['t-class'];
-  };
-  /**
    * 步骤条方向，有两种：横向和纵向
    * @default horizontal
    */
@@ -51,20 +44,12 @@ export interface TdStepsProps {
     value?: boolean;
   };
   /**
-   * 步骤条分割符
-   * @default line
+   * 步骤条顺序
+   * @default positive
    */
-  separator?: {
+  sequence?: {
     type: StringConstructor;
-    value?: 'line' | 'dashed' | 'arrow';
-  };
-  /**
-   * 自定义组件样式
-   * @default ''
-   */
-  style?: {
-    type: StringConstructor;
-    value?: string;
+    value?: 'positive' | 'reverse';
   };
   /**
    * 步骤条风格
@@ -74,68 +59,4 @@ export interface TdStepsProps {
     type: StringConstructor;
     value?: 'default' | 'dot';
   };
-}
-
-export interface TdStepItemProps {
-  /**
-   * 步骤描述
-   * @default ''
-   */
-  content?: {
-    type: StringConstructor;
-    value?: string;
-  };
-  /**
-   * 组件类名，用于设置组件外层元素元素类名
-   */
-  externalClasses?: {
-    type: ArrayConstructor;
-    value?: ['t-class', 't-class-content', 't-class-title', 't-class-description', 't-class-extra'];
-  };
-  /**
-   * 图标。传入 slot 代表使用插槽，其他字符串代表使用内置图标
-   */
-  icon?: {
-    type: StringConstructor;
-    value?: string;
-  };
-  /**
-   * 当前步骤的状态：默认状态（未开始）、进行中状态、完成状态、错误状态
-   * @default default
-   */
-  status?: {
-    type: StringConstructor;
-    value?: StepStatus;
-  };
-  /**
-   * 自定义组件样式
-   * @default ''
-   */
-  style?: {
-    type: StringConstructor;
-    value?: string;
-  };
-  /**
-   * 子步骤条，仅支持 layout  = 'vertical' 时
-   * @default []
-   */
-  subStepItems?: {
-    type: ArrayConstructor;
-    value?: SubStepItem[];
-  };
-  /**
-   * 标题
-   * @default ''
-   */
-  title?: {
-    type: StringConstructor;
-    value?: string;
-  };
-}
-
-export type StepStatus = 'default' | 'process' | 'finish' | 'error';
-
-export interface SubStepItem {
-  status: StepStatus;
-  title: string;
 }
